+++ packages/local/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.33](https://gitee.com/newgateway/vtj/compare/@vtj/local@0.7.32...@vtj/local@0.7.33) (2024-04-08)
+
+**Note:** Version bump only for package @vtj/local
+
+
+
+
+
+
 ## [0.7.32](https://gitee.com/newgateway/vtj/compare/@vtj/local@0.7.31...@vtj/local@0.7.32) (2024-04-03)
 
 **Note:** Version bump only for package @vtj/local